# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tax API",
        "title": "Tax API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "paths": {
        "/tax/calculations": {
            "post": {
                "description": "Calculate Tax",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "Calculate Tax",
                "parameters": [
                    {
                        "description": "Body for calculation request",
                        "name": "CalculationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tax.CalculationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tax.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tax.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tax.Err"
                        }
                    }
                }
            }
        },
        "/tax/calculations/upload-csv": {
            "post": {
                "description": "Calculate Tax for upload CSV file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "summary": "Calculate Tax for upload CSV file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Uploaded CSV for tax calculation",
                        "name": "taxes.csv",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tax.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tax.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tax.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "tax.AllowanceRequest": {
            "type": "object",
            "properties": {
                "allowanceType": {
                    "type": "string",
                    "example": "donation"
                },
                "amount": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "tax.CalculationRequest": {
            "type": "object",
            "properties": {
                "allowances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tax.AllowanceRequest"
                    }
                },
                "totalIncome": {
                    "type": "number",
                    "example": 500000
                },
                "wht": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "tax.Err": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "tax.Response": {
            "type": "object",
            "properties": {
                "tax": {
                    "type": "number",
                    "example": 29000
                },
                "taxLevel": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tax.TaxLevelResponse"
                    }
                }
            }
        },
        "tax.TaxLevelResponse": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string",
                    "example": "0-150,000"
                },
                "tax": {
                    "type": "number",
                    "example": 0
                }
            }
        }
    }
}